+++ Receivers/linux/pulseaudio/README.md
@@ -20,9 +20,18 @@ Run `make` command.
 $ scream-pulse
 ```
 
+This starts the Scream client in multicast mode.
 If your machine has more than one network interface, you may need to
 set the interface name which receives scream packets.
 
 ```shell
 $ scream-pulse -i eth0
+```
+
+### IVSHMEM (Shared memory) mode
+
+Make sure to have read permission for the shared memory device and execute
+
+```shell
+$ scream-pulse -m /dev/shm/scream-ivshmem
 ```
