--- Receivers/pulseaudio-ivshmem/Makefile
@@ -1,5 +1,0 @@
-CC=gcc
-CFLAGS = -Wall -O2
-
-scream-pulse: scream-ivshmem-pulse.c
-	$(CC) $(CFLAGS) -o scream-ivshmem-pulse scream-ivshmem-pulse.c -lpulse-simple -lpulse
